requirements.txt
streamlit
langchain[docarray]
langchain-community
openai
docarray
PyPDF2
python-dotenv
unstructured
psutil
unstructured[pdf]
tiktoken
